#65 패키지 변경 - 상태를 나타낼 enum class 분리, ViewModelFactory 패키지 이동
