+++ app/src/main/java/com/example/woowagithubrepositoryapp/ui/common/states.kt
@@ -1,0 +1,17 @@
+package com.example.woowagithubrepositoryapp.ui.common
+
+enum class DataLoading {
+    BEFORE, NOW, AFTER
+}
+
+enum class Tab(val text : String) {
+    ISSUE("Issue"), NOTI("Notifications")
+}
+
+enum class IssueLoadType{
+    CREATE, LOAD, PAGING
+}
+
+enum class SearchType {
+    CREATE, SEARCH
+}
