+++ app/src/main/java/com/example/woowagithubrepositoryapp/utils/Constants.kt
@@ -5,16 +5,4 @@ import com.example.woowagithubrepositoryapp.BuildConfig
 object Constants {
     const val GITHUB_CLIENT_ID = BuildConfig.githubClientId
     const val GITHUB_CLIENT_SECRET = BuildConfig.githubClientSecret
-
-    enum class DataLoading {
-        BEFORE, NOW, AFTER
-    }
-
-    enum class Tab(val text : String) {
-        ISSUE("Issue"), NOTI("Notifications")
-    }
-
-    enum class IssueLoadType{
-        CREATE, LOAD, PAGING
-    }
 }
